Fixed memory management issues - added parent view controller (UINavigationController) - added some colors in the table cells

+++ Fetcher/DayData.h
@@ -11,12 +11,12 @@
 @interface DayData : NSObject
 
 @property(nonatomic,strong,readonly) NSDate* date;
-@property(nonatomic,assign,readonly) uint open;
-@property(nonatomic,assign,readonly) uint high;
-@property(nonatomic,assign,readonly) uint low;
-@property(nonatomic,assign,readonly) uint close;
-@property(nonatomic,assign,readonly) uint volume;
-@property(nonatomic,assign,readonly) uint adjusted_close;
+@property(nonatomic,assign,readonly) NSUInteger open;
+@property(nonatomic,assign,readonly) NSUInteger high;
+@property(nonatomic,assign,readonly) NSUInteger low;
+@property(nonatomic,assign,readonly) NSUInteger close;
+@property(nonatomic,assign,readonly) NSUInteger volume;
+@property(nonatomic,assign,readonly) NSUInteger adjusted_close;
 
 -(id) initWithScanner:(NSScanner*) scanner;
